// src/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  CARE_REQUEST_ASSIGNED = 'CARE_REQUEST_ASSIGNED',
  CARE_REQUEST_ACCEPTED = 'CARE_REQUEST_ACCEPTED',
  CARE_REQUEST_DECLINED = 'CARE_REQUEST_DECLINED',
  CARE_REQUEST_REASSIGNED = 'CARE_REQUEST_REASSIGNED',
  CARE_REQUEST_CANCELLED = 'CARE_REQUEST_CANCELLED',
  CARE_APPOINTMENT_SCHEDULED = 'CARE_APPOINTMENT_SCHEDULED',
  CARE_APPOINTMENT_CANCELLED = 'CARE_APPOINTMENT_CANCELLED',
  PROVIDER_ONBOARDING_SUBMITTED = 'PROVIDER_ONBOARDING_SUBMITTED',
  PROVIDER_APPROVED = 'PROVIDER_APPROVED',
  PROVIDER_REJECTED = 'PROVIDER_REJECTED',
}